+++ include/swift/SIL/SILModule.h
@@ -21,6 +21,7 @@
 #include "swift/AST/Builtins.h"
 #include "swift/AST/Module.h"
 #include "swift/AST/SILOptions.h"
+#include "swift/AST/SILLayout.h"
 #include "swift/Basic/LangOptions.h"
 #include "swift/Basic/Range.h"
 #include "swift/SIL/SILCoverageMap.h"
@@ -29,7 +30,6 @@
 #include "swift/SIL/SILFunction.h"
 #include "swift/SIL/SILGlobalVariable.h"
 #include "swift/SIL/Notifications.h"
-#include "swift/SIL/SILLayout.h"
 #include "swift/SIL/SILType.h"
 #include "swift/SIL/SILVTable.h"
 #include "swift/SIL/SILWitnessTable.h"
