+++ lib/SIL/CMakeLists.txt
@@ -21,7 +21,6 @@ add_swift_library(swiftSIL STATIC
   SILGlobalVariable.cpp
   SILInstruction.cpp
   SILInstructions.cpp
-  SILLayout.cpp
   SILLocation.cpp
   SILModule.cpp
   SILOpenedArchetypesTracker.cpp
